requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
python-dotenv>=0.19.0
PyPDF2>=3.0.0
python-pptx>=0.6.0
langchain==0.0.350
openai==1.3.7
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary>=2.9.0  # PostgreSQL adapter 
PyPDF2
python-pptx
python-docx
